cluster: Move a function to gateway_init.go
The addDefaultConntrackRules() is used for gateways only
and makes sense to move to gateway specific file.
